v1.6.0 for package rename and update to 6-sanpshot and Spark 2.1.3
