+++ README.rst
@@ -11,7 +11,7 @@ Getting Started
 
 Prerequisites
 --------------
-CDAP version 3.5.x or higher.
+CDAP version 6.0.0-SNAPSHOT or higher.
 
 Building Plugins
 ----------------
@@ -58,7 +58,7 @@ CDAP IRC Channel: #cdap on irc.freenode.net
 License and Trademarks
 ======================
 
-Copyright © 2015-2016 Cask Data, Inc.
+Copyright © 2015-2019 Cask Data, Inc.
 
 Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
 in compliance with the License. You may obtain a copy of the License at
@@ -74,5 +74,3 @@ Cask is a trademark of Cask Data, Inc. All rights reserved.
 
 Apache, Apache HBase, and HBase are trademarks of The Apache Software Foundation. Used with
 permission. No endorsement by The Apache Software Foundation is implied by the use of these marks.
-
-.. |(Hydrator)| image:: http://cask.co/wp-content/uploads/hydrator_logo_cdap1.png
